Separate Trinkets-related code into conditionally loaded classes, instead of mixing into ourselves
this is much cleaner :)
