+++ src/main/java/io/github/alabasteralibi/simplyboots/SimplyBootsTrinketsDelegate.java
@@ -1,11 +1,10 @@
-package io.github.alabasteralibi.simplyboots.mixins;
+package io.github.alabasteralibi.simplyboots;
 
 import com.google.common.collect.Multimap;
 import dev.emi.stepheightentityattribute.StepHeightEntityAttributeMain;
 import dev.emi.trinkets.api.SlotReference;
 import dev.emi.trinkets.api.Trinket;
 import dev.emi.trinkets.api.TrinketsApi;
-import io.github.alabasteralibi.simplyboots.registry.SimplyBootsItems;
 import io.github.alabasteralibi.simplyboots.registry.SimplyBootsTags;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
@@ -14,19 +13,13 @@ import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.UUID;
 
 import static io.github.alabasteralibi.simplyboots.items.BootItem.STEP_BOOST_MODIFIER;
 
-@Mixin(SimplyBootsItems.class)
-public abstract class SimplyBootsTrinketMixin {
-    @Inject(method = "register", at = @At("TAIL"))
-    private static void registerTrinkets(String name, Item item, CallbackInfoReturnable<Item> cir) {
+public class SimplyBootsTrinketsDelegate {
+    public static void registerItem(Item item) {
         TrinketsApi.registerTrinket(item, new Trinket() {
             // Prevents equipping boots if a pair is already equipped as armor.
             @Override
